Save User

+++ src/main/java/com/mfs/merchantQR/repo/TblRoleRepo.java
@@ -3,5 +3,5 @@ package com.mfs.merchantQR.repo;
 import com.mfs.merchantQR.model.TblRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TblRoleRepo extends JpaRepository<TblRole,Long> {
+public interface TblRoleRepo extends JpaRepository<TblRole,Integer> {
 }
